requirements.txt
pandas
fastparquet
pyarrow
huggingface_hub
python-dotenv

# Model training dependencies
torch
transformers
scikit-learn
tqdm
datasets

# For synthetic data generation (optional)
langchain-openai
portkey-ai